↺ merge svn merge -r 1734:1772 https://amethyst.yui.ne.jp/svn-dev/pc98/np2/trunk svn merge -r 1776:1777 https://amethyst.yui.ne.jp/svn-dev/pc98/np2/trunk svn merge -r 1764:1783 http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui/WORK_01 svn merge -r 1792:1918 https://amethyst.yui.ne.jp/svn-dev/pc98/np2/trunk
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@1921 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ reform_ui/sdl2/MacOSX/np2sdl2/misc/usbdev.cpp
@@ -329,7 +329,7 @@ void CUsbDev::Close()
  * @param[in] cbBuffer バッファ長
  * @return サイズ
  */
-int CUsbDev::CtrlXfer(int nType, int nRequest, int nValue, int nIndex, void* lpBuffer, size_t cbBuffer)
+int CUsbDev::CtrlXfer(int nType, int nRequest, int nValue, int nIndex, void* lpBuffer, int cbBuffer)
 {
 	if (m_interface == NULL)
 	{
@@ -360,7 +360,7 @@ int CUsbDev::CtrlXfer(int nType, int nRequest, int nValue, int nIndex, void* lpB
  * @param[in] cbBuffer バッファ長
  * @return サイズ
  */
-int CUsbDev::WriteBulk(const void* lpBuffer, size_t cbBuffer)
+int CUsbDev::WriteBulk(const void* lpBuffer, int cbBuffer)
 {
 	if (m_interface == NULL)
 	{
@@ -382,7 +382,7 @@ int CUsbDev::WriteBulk(const void* lpBuffer, size_t cbBuffer)
  * @param[in] cbBuffer バッファ長
  * @return サイズ
  */
-int CUsbDev::ReadBulk(void* lpBuffer, size_t cbBuffer)
+int CUsbDev::ReadBulk(void* lpBuffer, int cbBuffer)
 {
 	if (m_interface == NULL)
 	{
